Implement external stBTC shares minting
Here we implement a possibility of minting stBTC tokens by external contract.
The feature will be used by Mezo Portal to mint stBTC tokens for users who
deposited tBTC and WBTC there.

Product Pitch: https://coda.io/d/Mezo_d_vIDNPK008/2-Drafts-Ready-for-Review_su1QZ#Draft-Product-Pitches_tuH_B/r16

\### Vault Equation

The ERC4626 is a vault that mints non-rebasing stBTC tokens, which means
there is a conversion rate between the assets (tBTC) and the shares (stBTC).
The vault equation is as follows:
```
sharesToReceive * totalAssets = assetsToDeposit * totalSupply
```

Any change of the total stBTC supply has to be reflected in the total assets
value to keep the equation balanced. The total assets value is the sum of
the assets deposited by users and the assets reflecting the debt minted
by the external contract.

The `totalAssets` function has been updated to include the taken debt that
was not covered with the deposited assets.

\### Debt Allowance

The governance can set a debt allowance for the external contract. The debt
allowance is the maximum amount of assets that the external contract can
take the debt for. The debt allowance is set in the tBTC token units.

\### Debt Minting
The external contract can mint stBTC tokens by taking the debt for the
assets that correspond to the minted shares but not deposited in the vault.
The debt minting is limited by the debt allowance.

The external contract can specify the receiver of the minted stBTC tokens.
The receiver can be the external contract itself or any other address.
Regardless of the receiver the debt will be assigned to the external contract
calling the mint function.

\### Debt Cancellation

To cancel the debt the external contract has to call the `cancelDebt` function.
The function expects the stBTC tokens allowance from the external contract
to the vault, to be able to burn the shares.

The amount of shares to burn is calculated based on the debt that the external
contract wants to cancel and the conversion rate between the shares and the
assets. The `previewCancelDebt` function can be used to calculate the amount
of shares required to cancel the given amount of debt assets.

\### Yield and Loss

In the vault that doesn't generate any yield or losses the conversion rate
of stBTC and tBTC is always 1:1. Meaning that when the external contract
mints a debt of 10 BTC they will receive 10 stBTC and will have to return
10 stBTC to cancel the debt.

In the case of the vault generating yield or losses the conversion rate
will be different than 1:1. The external contract will have to return more
or less stBTC than the minted debt to cancel the debt.

\### Non-fungible withdrawals

The changes introduced here depend on the [Non-fungible withdrawals](https://github.com/thesis/acre/pull/499).
It is a requirement to protect direct Acre depositors from the debt minters
draining the vault. The vault will not allow the debt minters to withdraw
the assets deposited by the direct Acre depositors.