Update integration tests
Adding the exit fee had an effect on integration tests.
We include the fee in the BitcoinRedeemer tests as the amounts in tests
are directly related to the fee.
For MezoAllocator, we update the tests to use zero fee, as including the
fee seems not that important for these tests.