Improve QueryKeys  management
Let's use a Query Key factory per feature.  It's just a simple object with entries and functions that will produce query keys, which you can then use in your custom hooks. This gives me a lot of flexibility, as each level builds on top of another, but is still independently accessible. Additionally, it separates user-specific queries from global ones coming from contracts.

More info: https://borstch.com/blog/development/understanding-the-role-of-querykeys-in-effective-data-management-with-react-query-library