Mint the requested amount of shares in mintDebt
Instead of accepting the asset amount in `mintDebt` and `repayDebt`,
we now accept the number of shares and convert them to asset amount
internally that is taken as a debt. This change allows to mint shares
in 1:1 relation to the amount deposited in Mezo.

The change is important for Mezo but is equally important to any other
potential external minter with non-fungible deposits that will be
integrated in the future.

If the asset amount was accepted in `mintDebt` as before, external
integrations would not be able to mint stBTC in a 1:1 ratio with the
token deposited. It is enough the vault generates any yield for the ratio
to not hold, with a donation from a third party included. This is not
problematic when deposits are fungible as eventually in Acre. Users spend
all their shares and redeem the corresponding amount of assets. When the
deposit is non-fungible, the user needs to have a precise amount of shares
to redeem their deposit. When they have less stBTC, some asset dust may
be abandoned and locked forever. If they have more stBTC, they can not use
it anyway.

For this reason, a better UX is achieved when minting stBTC in 1:1 ratio
to the asset deposited. Yet, in order not to break Acre's invariant equation,
the debt amount has to be calculated properly, based on the current condition
of the vault. It may happen that the external contract minting 1 stBTC will
have to pay back more or less than 1 underlying asset and this has to be
considered when integrating with stBTC.