Improved documentation of updateDebtAllowance
Explained the current debt value is intentionally not checked to allow
the governance reduce the debt allowance in case the depositor becomes
risky or malicious.

The governance may want to lower the allowed debt or at least set it to
the current debt value without the risk of being front-runned with
`mintDebt` calls from that depositor. One could make an argument this
is already too late, but there is no point in handcuffing the governance
- there are scenarios when being able to lower the allowed debt could
be helpful.