Compatibility functions for Mezo IReceiptToken
Mezo `IReceiptToken` expects `mintReceipt` and `burnReceipt` functions to
be exposed. We need a generic interface at Mezo because we may one day
decide to introduce separate receipt tokens for non-Bitcoin assets. An
alternative option to introducing those two functions in stBTC was to
implement an adapter contract but that would greatly complicate `msg.sender`
situation and require more gas for execution.