Fix calculating the redemption key
In `handleRedemptionRequested` function we want to build the redemption key
from the redeemer output script not from the redeemer output script hash
(`keccak256(redeemerOutputScript)`) because this param is available in
the event data. But in the `handleSubmitRedemptionProofCall` the
redeemer output script hash is available meaning there is no need to
calculate `keccak256(redeemerOutputScriptHash)`).

Here we split the build redemption key function into two:
- `buildRedemptionKey` - it builds redemption key assuming we pass the
  output script hash,
- `buildRedemptionKeyFromRedeemerOutputScript` - it builds redemption
  key from redeemer output script.