Show skeleton view until data is fetched
Let's display the data to the user only when we are sure that it has been fetched. Activities are saved in redux, so it was necessary to add an additional state to control when data are fully fetched. User data that that use tanstack query needs to be executed only when the prerequisites are completed. This will allow us to easily check when queries are first called by useIsFetching.