Fix `test` script in subgraph workspace
The `graph test` command is looking for the `subgraph.yaml` file by
default and there is no way of specifying the path to the subgraph
manifest. As a workaround we build the subgraph manifest for testing
purposes and save it in `subgraph.yaml` file so we can run tests
correctly.