Simplify redirection after closing the modal window
Instead of passing the `closeModal` function, let's just pass the `navigateToOnClose` parameter. The redirection logic will be handled in one place. The`react-router-dom` package has been updated to version `6.26.0` because there was previously a problem with the useNavigate hooks. `useNavigate` caused the component to be re-rendered. This is probably related to this issue - https://github.com/remix-run/react-router/issues/7634