Add deployment script for stBTC and MezoAllocator contract upgrades
These are upgrade scripts for stBTC and MezoAllocator contracts that
will be executed on Sepolia and Mainnet.

The solution is using proxy upgrade helpers function that will
deploy implementation contract and prepare transaction that will be
executed by the governance to upgrade the implementation on the proxy
contract.

It requires changes from https://github.com/keep-network/hardhat-helpers/pull/54
which branch is set in package.json as version of the package.
We need to update the version of `keep-network/hardhat-helpers` dependency
once it is released.