Update the `TokenBalanceInput` component
We should not truncate the trailing zeros in the deposit amount input so aggressively. Best way to handle this is to use `inputValue` instead an `amount`. The `inputValue` value is of type string so we will not truncate zeros as it was done using `fixedPointNumberToString`. However, a `bigint` value will still be passed to the form values.