Check if withdrawal amount exceeds the deposit balance
Here we improve the withdrawal function to check if the withdrawal amount
exceeds the deposit balance. If the withdrawal amount exceeds the deposit
balance, the function should revert with an error message.
Before these changes, the withdrawal function reverted on
`depositBalance -= uint96(amount);` with the following error message:
```
panic code 0x11 (Arithmetic operation overflowed outside of an unchecked block)
```