Include unallocated MezoAllocator balance in withdraw
In MezoAllocator contract in the totalAssets function we include the
balance of the deposit made in Mezo Portal as well as the balance of
the MezoAllocator contract that hasn't been yet allocated (this can be
due to a donation or rewards).

This change is to include the unallocated balance in the withdraw
function so that the user can withdraw the unallocated balance as well.
If we're not withdrawing the unallocated balance, a last user to withdraw
may not be able to complete the withdrawal as the amount of tBTC that is
transferred to the stBTC contract includes only the funds withdrawn from
the Mezo Portal contract.