Fix rounding of shares to assets conversions in debt minting
Previously, the shares to assets conversion was rounding down, which could lead
to a situation where the minted debt was against the vault's interest.

Let's imagine a scenario where the vault has a 10% loss, meaning that each
share is worth 0.9 assets.
If a user mints 1 share, the debt registered due to rounding down would
be 0. This can be used as a griefing attack, where a user could mint shares
without registering any debt.

To fix the issue, we now round up the shares to assets conversion in mintDebt
function.

Additionally, for the repayDebt function, we keep rounding down the shares to assets
to avoid the opposite issue, where a user could repay more debt than they should.

It may happen that conversion in `repayDebt` returns `0`, and it should be
validated by the caller. Since, the interface used by the Mezo Portal contract
`burnReceipt` function doesn't consider the return value of `repayDebt`, we
are performing the validation in the `burnReceipt` function. For consistency
we added a similar validation in the `mintReceipt` function.