Disable disconnection from the wallet level
Unisat in version v1.4.5 added a `disconnect` method by this our dApp began to behave strangely. The `handleAccountsChanged` function runs even when a user disconnects from the dApp level. Then when the user is disconnected and the `onDisconnect` is called again unisat will open a wallet window. Let's disable the disconnect feature from the wallet level for now.