Implement MultiAssetVault
In this commit we implement the MultiAssetVault contract. This contract is
responsible for managing deposits and withdrawals of multiple assets. It is
designed to be used in addition to the stBTC vault, to allow users to deposit
assets other than tBTC.

The stBTC token will not be generated for the assets deposited in the
MultiAssetVault.

The deposits created in this vault are routed directly to the Mezo Portal.
Each deposit is associated with a unique identifier, generated by the Mezo
Portal.

Deposit can be withdrawn by providing the deposit ID. The deposit is
fully withdrawn from the Mezo Portal and the assets are transferred to the
user.

Partial withdrawals are not supported. The user must withdraw the full
deposit.