Make the scrollbar visible when the sidebar is open
The sidebar has a higher z-index than the modal. When the scrollbar appears for the modal window it is covered by the sidebar. Therefore, we should check when the scrollbar is visible and then set the position of the Sidebar correctly.

Note that we can not change the z-index for the Sidebar because the container for the modal window covers all window and will not allow a user to click on the element.