Add build job to upload source maps to Sentry
The code merged to main branch will trigger a deployment on Netlify side.
We want to also upload the source maps to Sentry to enable source map support.
We don't want to upload the source maps for commits pushed to other branches,
so we set the SENTRY env variables only for push event which is triggered
on main branch. If SENTRY env variables are not set, the upload to Sentry
won't be invoked.