Revert "Add build job to upload source maps to Sentry"
This reverts commit 53fb8408540077caf5b3c5130de7126c5643e21a.

We need to upload the source maps to Sentry from the version that is
being deployed. The GitHub Actions build job is running for the main branch,
only to verify the build. But the dapp is build and deployed from Netlify.
We remove the upload step from the GitHub Actions, and we will add it to
Netlify build setup.