Update the BitcoinProvider interface
Add `signWithdrawMessage` function to support the clear signing with
Ledger Live. This function is optional so if the bitcoin provider
supports the clear signing it should implement this method and this
function will be used in the withdrawal flow.

Implement custom bitcoin provider and connector

We need to use the custom Acre module to support the clear signing.

Update Ledger Live manifests

We use new methods from the Ledger Live Wallet API and we need to set
them in manifest file.

Update orangekit packages

Adjust the Acre SDK to the new orangekit API.

Remove workaround for the session mechanism

We can sign the message with correct derivation path in Ledger Live.

Fix linting error