Fetch Mezo mats data from Acre API
In https://github.com/thesis/acre-bots/pull/49 we exposed the Mezo mats
data from the Acre API, so the dApp doesn't have to call the Mezo Portal
API anymore.
The data are cached in the Acre API, so the dApp doesn't have to worry
about hitting the rate limit of the Mezo Portal API (if introduced in the future).
Additionally the dApp doesn't have to manage the Mezo API KEY anymore.