Update `useWallet` hook
The wagmi returns the connector object even if we call `disconnect`
function. In that case the connector has basic information like: id,
name, type and uuid and the status is set to `connected`. So here we
need to make sure the `getAccounts` and `getBitcoinAddress` functions
exist in the connector object.