Add `bignumber.js` package to dapp
Add the `bignumber.js` lib used by the Ledger Live packages. When
ignoring the TS error we're losing the advantage of the type safety TS
provides. We assume that the external library performs the conversion,
but if the implementation in the Ledger library changes, we will only
find out at runtime. Here we pass the satoshi value as a BigNumber
object to the Ledger Live module.