Fix wallet connection bug
The bug is that the `window.unisat` and
`window.XverseProviders.BitcoinProvider` have `undefined` values. The
orangekit package refers to these fields under the hood and the
`isInstalled` function always returns `false`. Due to this, users were
redirected to download the wallet extension even if they had it
installed.

To fix this bug here we use the dynamic import to get functions that
create connectors.