Set chain id based on the env variable
Define the chain id based on the env variable and use it in the
`connect` function from wagmi. We have separate builds for each network
and we do not change chain dynamically so we can use hardcoded value in
`connect` function. If we use the `useChainId` hook from wagmi it will
return the cached chain id. So when for example we were locally testing
a dapp on the testnet and then changed the env variables to integrate
with the mainnet, the wagmi used the cached chain id for the testnet and
we couldn't connect to the wallet.