Update the referral mechanism
Detect the referral code in the main router loader instead of the hooks.
Using a router loader is simpler and we avoid race between two hooks
`useDetectEmbed` and `useDetectReferral`.

We noticed a bug with hooks implementation - `acre.referral` set to `0`
for Ledger Live after clearing local storage. The `acre.referral` key
should always be set to `2083` value for users using Acre in Ledger
Live.