Enable Sentry ExtraErrorData integration
Enable [ExtraErrorData](https://docs.sentry.io/platforms/javascript/guides/react/configuration/integrations/extraerrordata/)
integration in Sentry.

This integration serialized extra data attached to the error event. It is
useful to attach additional context to the error event, as currently the
objects are not serialized and are visible as `[Object]` in the Sentry dashboard.