Update estimated time for processing of small deposits from 1 hour to 2 hours
Deposit duration for small deposits was increased from 1 hour to 2 hours.
This change was made to reflect the actual time it takes for small deposits
to be processed. We observed that small deposits are taking in general 1.5 hours
to finalize, so we decided to round up for better user experience.