Support direct navigation in Ledger Live App
To ensure users are directed to the correct LiveApp page for specific
actions (such as staking/unstaking) triggered from Ledger Live, the Acre
dapp must implement URL parameters to support direct navigation.

We want to connect to the account passed via URL. The Ledger Live App
passes the `accountId` parameter in URL when redirecting to Acre dapp.
This parameter has the following pattern
`js:2:bitcoin_testnet:<xpub>:<address_type>`. To connect to a given
account we need to get the extended public key from this parameter.

To connect to a given account by xpub we define the new option in the
provider: `tryConnectToAccountByXpub`. The Ledger Live Bitcoin provider
will try to find this account and connect.