Move dapp tests to the root test folder
For consistency with other packages and projects we are moving the dapp
tests to the root test folder.

There are many possible approaches to organize the tests, but here we
are following the same structure as the other packages.
If we need to change this in the future, we can discuss and find the best
approach.